get rid of models fix get reminder
