+++ aifunctions/get_stock_price_function.json
