--- models/models.go
@@ -1,9 +1,0 @@
-package models
-
-// used for sending a message on a specific discord channel
-type ChannelMessage struct {
-	Message     string `json:"message"`
-	TimerLength int    `json:"timer_length,omitempty"`
-	ChannelID   string `json:"channel_id,omitempty"`
-	UserID      string `json:"user_id,omitempty"`
-}
